deepseek_v41: match KV dtype to the queries before fused attention
The sliding-window KV cache is stored float32 while queries carry the model
dtype, so `mx.fast.scaled_dot_product_attention` received mismatched inputs and
fell off its fused path. Only the two ratio-0 layers reach that call -- every
other layer takes the sparse pooled path -- but those two calls cost about 79 ms
each against 0.1 ms for the same shapes with matching dtypes.

The cast is lossless: the cache holds values that already passed through
`fake_quant_fp8_ue8m0`, so each is an fp8 mantissa scaled by a power of two and
both fit the narrower type exactly. Sampled tokens are unchanged.

Measured on DeepSeek-V4.1-Flash-MLX-mixed-4_8bit (M3 Ultra, 512 GB):

  decode step   207.1 -> 48.9 ms    4.82 -> 20.45 tok/s   4.24x

The layers isolate it cleanly: layers 2..39 cost ~1.0 ms each (38 of them =
46.8 ms), while adding either ratio-0 layer added ~155 ms.

+++ mlx_vlm/models/deepseek_v41/language.py
@@ -672,7 +672,12 @@ class DeepseekV41Attention(nn.Module):
                 )
         if out is None:
             mask = window_mask
-            kv = window_kv[:, None]
+            # The window cache is float32 while queries are the model dtype, and
+            # mixed dtypes drop this call off the fused attention path: measured
+            # 79 ms against 0.1 ms for one decode step's call. The cast is
+            # lossless because the cache holds fake-quantized fp8 values, whose
+            # mantissa and exponent both fit the narrower type.
+            kv = window_kv[:, None].astype(q.dtype)
             out = mx.fast.scaled_dot_product_attention(
                 q,
                 kv,
